Dynamic Docker image tagging with commit SHA
Added a step to set the image tag using the first seven
characters of the Git commit SHA, stored in the
`IMAGE_TAG` environment variable. Updated Docker build
and push commands to utilize this dynamic tag instead
of a static build number. Modified deployment to use
the new image tag instead of the hardcoded 'latest' tag.